csp_print: Add option to toggle csp_print
csp_print() is now configurable with the CSP_ENABLE_CSP_PRINT option. This
commit fixes the mixed usage of CSP_HAVE_STDIO, which is used to both
enable / disable csp_print() and for snprintf() / sscanf() conditional
compilation.

With this commit, CSP_ENABLE_CSP_PRINT and CSP_HAVE_STDIO are orthogonal,
you can enable or disable either or both without interfering each other.

  - CSP_ENABLE_CSP_PRINT = 1 enables csp_print().
  - CSP_HAVE_STDIO = 1 enables sscanf() and snprintf() related functions.

You can cleanly disable all stdio.h related code with CSP_HAVE_STDIO = 0.

A few notes for minor changes:

  - csp_rtable_stdio.c is compiled when and only when CSP_HAVE_STDIO = 1.
  - csp_print_func is changed depending on CSP_HAVE_STDIO.
  - csp_conn_t conn is marked as `unused` because csp_print() might be noop.

With this commit, csp_conn.c, csp_rtable_stdio, and csp_debug.c are the
only source file with `#include <stdio.h>`.
