+++ include/csp/csp_iflist.h
@@ -14,8 +14,13 @@ csp_iface_t * csp_iflist_get_by_name(const char *name);
 csp_iface_t * csp_iflist_get_by_addr(uint16_t addr);
 csp_iface_t * csp_iflist_get_by_subnet(uint16_t addr);
 
-void csp_iflist_print(void);
 csp_iface_t * csp_iflist_get(void);
 
 /* Convert bytes to readable string */
 unsigned long csp_bytesize(unsigned long bytes, char *postfix);
+
+#if (CSP_ENABLE_CSP_PRINT)
+void csp_iflist_print(void);
+#else
+inline void csp_iflist_print(void) {}
+#endif
