+++ meson_options.txt
@@ -7,6 +7,7 @@ option('enable_python3_bindings', type: 'boolean', value: false, description: 'B
 
 option('version', type: 'integer', value: 1, description: 'Which version of CSP to use.')
 option('packet_padding_bytes', type: 'integer', value: 8, description: 'Number of bytes to include before the packet data (must be minimum 8)')
+option('enable_csp_print', type: 'boolean', value: true, description: 'Enable csp_print()')
 option('have_stdio', type: 'boolean', value: true, description: 'Use print and scan functions (some features may be missing without)')
 
 # Memory tuning parameters:
