+++ src/csp_debug.c
@@ -11,6 +11,7 @@ uint8_t csp_dbg_inval_reply;
 uint8_t csp_dbg_rdp_print;
 uint8_t csp_dbg_packet_print;
 
+#if (CSP_ENABLE_CSP_PRINT)
 #if (CSP_HAVE_STDIO)
 #include <stdarg.h>
 #include <stdio.h>
@@ -20,4 +21,7 @@ __attribute__((weak)) void csp_print_func(const char * fmt, ...) {
     vprintf(fmt, args);
     va_end(args);
 }
+#else
+__attribute__((weak)) void csp_print_func(const char * fmt, ...) {}
+#endif
 #endif
