+++ src/csp_hex_dump.c
@@ -4,7 +4,7 @@
 #include <csp/csp_debug.h>
 #include <stddef.h>
 
-#if (CSP_HAVE_STDIO)
+#if (CSP_ENABLE_CSP_PRINT)
 void csp_hex_dump_format(const char * desc, void * addr, int len, int format) {
 	int i;
 	unsigned char buff[17];
