+++ src/meson.build
@@ -16,7 +16,6 @@ csp_sources += files([
 	'csp_qfifo.c',
 	'csp_route.c',
 	'csp_rtable_cidr.c',
-	'csp_rtable_stdio.c',
 	'csp_service_handler.c',
 	'csp_services.c',
 	'csp_id.c',
@@ -25,6 +24,10 @@ csp_sources += files([
 
 if yaml_dep.found()
 	csp_sources += files('csp_yaml.c')
+endif
+
+if get_option('have_stdio')
+	csp_sources += files('csp_rtable_stdio.c')
 endif
 
 subdir('arch')
